Update 점 찍기.java

+++ 프로그래머스/2/140107. 점 찍기/점 찍기.java
@@ -1,10 +1,10 @@
 class Solution {
-    public long solution(long k, long d) {
+    public long solution(int k, int d) {
         long answer = 0;
 
-        for(long x = 0; x <= d; x += k){
-            long maxY = (long) Math.sqrt(d*d - x*x); 
-            answer += maxY / k + 1;
+        for (int x = 0; x <= d; x += k){
+            long maxY = (long) Math.sqrt((long)d*d - (long)x*x); 
+            answer += (maxY / k) + 1;
         }
         
         return answer;
